refactor(tui): split shared sheet controls
