+++ station/src/station/view/settings/WidgetSettingsPanelView.tsx
@@ -7,10 +7,10 @@ import type {
 } from "@station/dashboard-core/state";
 import { widgetSettingsPanelModel } from "@station/dashboard-core/selectors";
 import type { WidgetSettingsItem } from "@station/dashboard-core/selectors";
+import { fit } from "../controls/sheetText.js";
 import { SemanticScrollRegion } from "../layout/scroll/SemanticScrollViewport.js";
 import { semanticItemRenderableId } from "../layout/scroll/scrollViewport.js";
 import { widgetSettingsFrame } from "../layout/settingsPanelFrame.js";
-import { fit } from "../sheets/parts.js";
 import { toOpenTuiColor, toOpenTuiOpaqueColor, useStationTheme } from "../../../theme/index.js";
 import {
   stationMouseProps,
