+++ station/src/station/view/sheets/BottomSheetFrameView.test.tsx
@@ -2,10 +2,11 @@ import { afterEach, describe, expect, it } from "bun:test";
 import { testRender } from "@opentui/react/test-utils";
 import { act, type Dispatch, type SetStateAction, useState } from "react";
 import { nativeStationTheme, StationThemeProvider } from "../../../theme/index.js";
+import { SheetChoiceLine } from "../controls/sheetPicker.js";
+import { SheetFooter, SheetLine } from "../controls/sheetText.js";
 import { semanticItemRenderableId } from "../layout/scroll/scrollViewport.js";
 import { StationHoverProvider, StationMouseProvider } from "../stationMouseContext.js";
 import { BottomSheetFrameView } from "./BottomSheetFrameView.js";
-import { SheetChoiceLine, SheetFooter, SheetLine } from "./parts.js";
 
 const teardowns: Array<() => void> = [];
 afterEach(async () => {
