+++ station/src/station/view/sheets/CreateGroupSheetView.tsx
@@ -1,10 +1,11 @@
 import { createGroupSheetContent } from "@station/dashboard-core/selectors";
 import { cellWidth } from "@station/dashboard-core/text";
 import type { DashboardScreenView } from "@station/dashboard-core/state";
+import { SheetButtonRow, SheetControlRow } from "../controls/sheetButtons.js";
+import { SheetFooter } from "../controls/sheetText.js";
 import { EditableTextInputView } from "../EditableTextInputView.js";
 import { bottomSheetContentWidth } from "../layout/bottomSheetFrame.js";
 import { BottomSheetFrameView } from "./BottomSheetFrameView.js";
-import { SheetButtonRow, SheetControlRow, SheetFooter } from "./parts.js";
 
 export type CreateGroupSheetViewProps = {
   screen: Extract<DashboardScreenView, { name: "createGroup" }>;
