+++ station/src/station/view/sheets/ProjectChoiceSheetView.tsx
@@ -2,9 +2,10 @@ import type { ProjectId } from "@station/contracts";
 import { selectProjectChooserChoices } from "@station/dashboard-core/selectors";
 import type { DashboardSnapshotView, DashboardStateView } from "@station/dashboard-core/state";
 import { providerHealthColor, useStationTheme } from "../../../theme/index.js";
+import { SheetChoiceLine } from "../controls/sheetPicker.js";
+import { SheetFooter } from "../controls/sheetText.js";
 import { bottomSheetContentWidth } from "../layout/bottomSheetFrame.js";
 import { BottomSheetFrameView } from "./BottomSheetFrameView.js";
-import { SheetChoiceLine, SheetFooter } from "./parts.js";
 
 export type ProjectChooserMode = "projectCollapse" | "projectSettingsPicker";
