+++ station/src/station/view/sheets/RenameSessionSheetView.tsx
@@ -1,11 +1,12 @@
 // OpenTUI port of apps/tui's RenameSessionBottomSheet.
 import { truncateCells } from "@station/dashboard-core/text";
 import type { DashboardScreenView } from "@station/dashboard-core/state";
-import { EditableTextInputView } from "../EditableTextInputView.js";
 import { toOpenTuiColor, useStationTheme } from "../../../theme/index.js";
+import { SheetButtonRow } from "../controls/sheetButtons.js";
+import { SheetFooter, SheetLabelValue, SheetText } from "../controls/sheetText.js";
+import { EditableTextInputView } from "../EditableTextInputView.js";
 import { bottomSheetContentWidth } from "../layout/bottomSheetFrame.js";
 import { BottomSheetFrameView } from "./BottomSheetFrameView.js";
-import { SheetButtonRow, SheetFooter, SheetLabelValue, SheetText } from "./parts.js";
 
 export type RenameSessionSheetViewProps = {
   state: Extract<DashboardScreenView, { name: "renameSession"; step: "editName" }>;
